python/requirements.txt
pandas~=2.1.2
matplotlib~=3.7.1
opencv-python~=4.7.0
lxml~=4.9.2
GitPython~=3.1.31
swc-aeon @ git+https://github.com/SainsburyWellcomeCentre/aeon_api.git@v0.1.0